fix: resolve import paths after client/server test co-location refactor
