+++ src/client/pages/Editor/Pages/EditorHome.tsx
@@ -1,7 +1,7 @@
 import { User } from 'blaise-api-node-client';
 import AsyncContent from '../../shared/AsyncContent';
 import { Survey } from '../../../../common/interfaces/surveyInterface';
-import { getSurveys } from '../../api/NodeApi';
+import { getSurveys } from '../../../api/NodeApi';
 import { useAsyncRequestWithParam } from '../../../utils/useAsyncRequest';
 import SurveysList from '../../shared/SurveysList';
 import ErrorPanel from '../../shared/ErrorPanel';
