+++ src/client/pages/Supervisor/Pages/SupervisorHome.tsx
@@ -1,7 +1,7 @@
 import { User } from 'blaise-api-node-client';
 import { Survey } from '../../../../common/interfaces/surveyInterface';
 import { useAsyncRequestWithParam } from '../../../utils/useAsyncRequest';
-import { getSurveys } from '../../api/NodeApi';
+import { getSurveys } from '../../../api/NodeApi';
 import AsyncContent from '../../shared/AsyncContent';
 import SurveysList from '../../shared/SurveysList';
 import ErrorPanel from '../../shared/ErrorPanel';
