+++ src/server/surveyMapper.ts
@@ -1,4 +1,4 @@
-import { QuestionnaireDetails, Survey } from '../../common/interfaces/surveyInterface';
+import { QuestionnaireDetails, Survey } from '../common/interfaces/surveyInterface';
 
 export default function mapSurveys(questionnaires: QuestionnaireDetails[]): Survey[] {
   const surveys: Survey[] = [];
